Add missing child care variables
