+++ openfisca_aotearoa/tests/acc/social_rehab/child_care.yaml
@@ -1,0 +1,91 @@
+---
+- name: Child Care entitlement
+  period: 2019-02
+  input:
+    persons:
+      Claimant_Scot:
+        acc__has_a_covered_injury: true
+        acc__part_3__has_lodged_claim: true
+        acc__assessed_as_having_a_need_caused_by_this_covered_injury: true
+        acc__corporation_has_regard_to_provide_or_contribute_to_child_care: true
+        acc__key_aspect_is_necessary_and_appropriate: true
+        acc__key_aspect_is_of_the_quality_required_for_that_purpose: true
+        acc__is_present_in_nz: true
+      Child_Mary:
+        age: 3
+    families:
+      smiths:
+        parents: [Claimant_Scot]
+        children: [Child_Mary]
+  output:
+    persons:
+      Claimant_Scot:
+        acc__is_entitled_to_child_care: true
+
+# - name: Child Care - ACC decides no
+#   period: 2019-02
+#   input:
+#     acc__has_a_covered_injury: true
+#     acc__part_3__has_lodged_claim: true
+#     acc__assessed_as_having_a_need_caused_by_this_covered_injury: true
+#     acc__corporation_has_regard_to_provide_or_contribute_to_child_care: false
+#     acc__key_aspect_is_necessary_and_appropriate: true
+#     acc__key_aspect_is_of_the_quality_required_for_that_purpose: true
+#     acc__is_present_in_nz: true
+#   output:
+#     acc__is_entitled_to_child_care: false
+
+# - name: Child Care - but there's no injury
+#   period: 2019-02
+#   input:
+#     acc__has_a_covered_injury: false
+#     acc__part_3__has_lodged_claim: true
+#     acc__assessed_as_having_a_need_caused_by_this_covered_injury: true
+#     acc__corporation_has_regard_to_provide_or_contribute_to_child_care: true
+#     acc__key_aspect_is_necessary_and_appropriate: true
+#     acc__key_aspect_is_of_the_quality_required_for_that_purpose: true
+#     acc__is_present_in_nz: true
+#   output:
+#     acc__is_entitled_to_child_care: false
+
+# - name: Child Care - but they left NZ
+#   period: 2019-02
+#   input:
+#     acc__has_a_covered_injury: true
+#     acc__part_3__has_lodged_claim: true
+#     acc__assessed_as_having_a_need_caused_by_this_covered_injury: true
+#     acc__corporation_has_regard_to_provide_or_contribute_to_child_care: true
+#     acc__key_aspect_is_necessary_and_appropriate: true
+#     acc__key_aspect_is_of_the_quality_required_for_that_purpose: true
+#     acc__number_of_days_outside_nz: 25
+#   output:
+#     acc__is_entitled_to_child_care: true
+
+
+# - name: Child Care - but they left NZ
+#   period: 2019-02
+#   input:
+#     acc__has_a_covered_injury: true
+#     acc__part_3__has_lodged_claim: true
+#     acc__assessed_as_having_a_need_caused_by_this_covered_injury: true
+#     acc__corporation_has_regard_to_provide_or_contribute_to_child_care: true
+#     acc__key_aspect_is_necessary_and_appropriate: true
+#     acc__key_aspect_is_of_the_quality_required_for_that_purpose: true
+#     acc__number_of_days_outside_nz: 31
+#   output:
+#     acc__is_entitled_to_child_care: false
+
+
+# - name: Child Care - they left NZ, but ACC okays this
+#   period: 2019-02
+#   input:
+#     acc__has_a_covered_injury: true
+#     acc__part_3__has_lodged_claim: true
+#     acc__assessed_as_having_a_need_caused_by_this_covered_injury: true
+#     acc__corporation_has_regard_to_provide_or_contribute_to_child_care: true
+#     acc__number_of_days_outside_nz: 31
+#     acc__the_corporation_exercised_discretion_for_attendant_care_as_per_section_68_3: true
+#     acc__key_aspect_is_necessary_and_appropriate: true
+#     acc__key_aspect_is_of_the_quality_required_for_that_purpose: true
+#   output:
+#     acc__is_entitled_to_child_care: true
